requirements.txt

streamlit
openai
psycopg2-binary
pgvector
nltk
transformers
youtube-transcript-api
yt-dlp
openai-whisper
pymupdf
torch
